% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.Tenv.R
\name{predict.Tenv}
\alias{predict.Tenv}
\title{Predict method for TRR and TPR Fits.}
\usage{
\method{predict}{Tenv}(object, newdata, ...)
}
\arguments{
\item{object}{Tenv object, result of \code{\link{TRR}} and \code{\link{TPR}}.}

\item{newdata}{The data to be used for prediction. It can be vector, matrix or tensor for \code{\link{TRR}} fits, and can be matrix
or tensor for \code{\link{TPR}} fits.}

\item{...}{Not used. Arguments passed to or from other methods.}
}
\description{
Predicted response based on Tenv object returned from \code{\link{TRR}} and \code{\link{TPR}} functions.
}
\examples{
data("bat")
Xn <- bat$Xn
Yn <- bat$Yn
fit <- TRR(Xn, Yn, method="standard")
predict(fit, Xn)
}
