% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/square.R
\docType{data}
\name{square}
\alias{square}
\title{Square simulated data}
\format{A list consisting of two components:
\describe{
 \item{x}{A \eqn{32 \times 32 \times 200} tensor, each matrix \code{x@data[,,i]} represents a predictor observation.}
 \item{y}{A \eqn{1 \times 200} matrix, each entry represents a response observation.}
}}
\usage{
data("square")
}
\description{
Simulated data under tensor predictor regression (TPR) model. Each response observation is univariate, and each predictor observation is a matrix.
}
\details{
The dataset is generated from the tensor predictor regression (TPR) model:
\deqn{Y_i = \langle\mathbf{B}, \mathbf{X}_i\rangle + \epsilon_i, \quad i = 1,\ldots, n,}
where \eqn{n=200} and the regression coefficient \eqn{\mathbf{B} \in R^{32\times 32}} is a given image with rank 2, which has a square pattern. All the elements of the coefficient matrix \eqn{\mathbf{B}} are either 0.1 or 1. To make the model conform to the envelope structure, we construct the envelope basis \eqn{\boldsymbol{\Gamma}_k} and the covariance matrices \eqn{\boldsymbol{\Sigma}_k, k=1,2}, of predictor \eqn{\mathbf{X}} as following. With the singular value decomposition of \eqn{\mathbf{B}}, namely \eqn{\mathbf{B} = \boldsymbol{\Gamma}_1 \boldsymbol{\Lambda} \boldsymbol{\Gamma}_2^{\top}}, we choose the envelope basis as \eqn{\boldsymbol{\Gamma}_k \in R^{32 \times 2}, k=1,2}. Then the envelope dimensions are \eqn{u_1 =  u_2 = 2}. We set matrices \eqn{\boldsymbol{\Omega}_k = \mathbf{I}_2}  and  \eqn{\boldsymbol{\Omega}_{0k} = 0.01\mathbf{I}_{30}}, \eqn{k=1,2}. Then we generate the covariance matrices \eqn{\boldsymbol{\Sigma}_k = \boldsymbol{\Gamma}_k\boldsymbol{\Omega}_k\boldsymbol{\Gamma}_k^{\top} + \boldsymbol{\Gamma}_{0k}\boldsymbol{\Omega}_{0k}\boldsymbol{\Gamma}_{0k}^{\top}}, followed by normalization with their Frobenius norms. The predictor \eqn{\mathbf{X}_i} is then generated from two-way tensor (matrix) normal distribution \eqn{TN(\mathbf{0}; \boldsymbol{\Sigma}_1, \boldsymbol{\Sigma}_2)}. And the error term \eqn{\epsilon_i} is generated from standard normal distribution.
}
\examples{
data("square")
## Coefficients
coeff <- square$coefficients
image(-coeff@data[, , 1], axes=TRUE, col = grey(seq(0, 1, length = 256)))
title('Coefficient matrix')
}
\keyword{datasets}
