\name{subspace}
\alias{subspace}

\title{
The distance between two subspaces \code{span(A)} and \code{span(B)}.
}
\description{
This function calculate the distance between two subspaces \code{span(A)} and \code{span(B)}.
}
\usage{
subspace(A, B)
}

\arguments{
  \item{A}{A \eqn{p}-by-\eqn{u} matrix.}
  \item{B}{A \eqn{p}-by-\eqn{u} matrix.}
}

\value{
Returns a distance metric that is between 0 and 1
}

