% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.Tenv.R
\name{summary.Tenv}
\alias{summary.Tenv}
\alias{print.summary.Tenv}
\title{Summarize method for Tenv object.
Summary method for object returned from \code{\link{TRR}} and \code{\link{TPR}} functions.}
\usage{
\method{summary}{Tenv}(object, ...)

\method{print}{summary.Tenv}(x, ...)
}
\arguments{
\item{object}{an object of class "Tenv", as from \code{\link{TPR}} or \code{\link{TRR}}.}

\item{...}{arguments to be passed to or from other methods.}

\item{x}{an object of class "summary.Tenv", usually, a result of a call to summary.Tenv.}
}
\value{
\code{object} with additional components
\describe{
 \item{n}{sample size}
 \item{xdim}{dimensions of predictor}
 \item{ydim}{dimensions of response}
 \item{mse}{mean squared error based on \code{residuals(object)}}
 \item{p_val}{only for object returned from \code{\link{TRR}}, p-value for coefficients}
 \item{se}{only for object returned from \code{\link{TRR}}, standard error for coefficients}
}
}
\description{
The p-value and standard error for coefficients are only calculated for the object returned from \code{\link{TRR}}.
}
\details{
print.summary.Tenv print.summary.lm gives a more readable format of sample size, dimensions of datasets, mse,
and additionally gives \code{p-val} and \code{se} if \code{object} is returend from \code{\link{TRR}}.
}
\examples{
data("bat")
Xn <- bat$Xn
Yn <- bat$Yn
fit <- TRR(Xn, Yn, method="standard")
##print summary
summary(fit)
}
\seealso{
\code{\link{Tenv_Pval}}
}
