% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TensPLS_fit.R
\name{TensPLS_fit-deprecated}
\alias{TensPLS_fit-deprecated}
\title{Tensor envelope partial least squares (PLS) regression}
\usage{
TensPLS_fit(x, y, SigX, u)
}
\arguments{
\item{x}{A predictor tensor of dimension \eqn{p_1\times \cdots \times p_m}.}

\item{y}{The response vector of dimension \eqn{r}.}

\item{u}{The dimension of envelope subspace, \eqn{u=(u_1,\cdots,u_m)}.}

\item{SigX}{A matrix lists \eqn{\boldsymbol{\Sigma}_k, k=1,\cdots, m}, which determines the estimation of covariance matrix \eqn{\boldsymbol{\Sigma}=\boldsymbol{\Sigma}_m \otimes \cdots \otimes \boldsymbol{\Sigma}_1}.}
}
\value{
\item{Gamma}{The estimation of factor matrix \eqn{W_k, k=1,\cdots,m}.}
  \item{PGamma}{The projection matrix \eqn{W_k(W_k'\boldsymbol{\Sigma}_k W_k)^{-1}W_k'\boldsymbol{\Sigma}_k, k=1,\cdots,m}.}
}
\description{
Estimates the factor matrix \eqn{W_k, k=1,\cdots,m} in tensor PLS algorithm for tensor predictor regression (TPR), see Zhang, X., & Li, L. (2017).
}
\references{
Zhang, X., & Li, L. (2017). Tensor Envelope Partial Least-Squares Regression. Technometrics, 59(4), 426-436.
}
\seealso{
\code{\link{TRES-deprecated}}
}
\keyword{internal}
