% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bat.R
\docType{data}
\name{bat}
\alias{bat}
\title{Bat simulated data}
\format{A list consisting of three components:
\describe{
 \item{x}{A \eqn{1 \times 20} matrix}
 \item{y}{A \eqn{64\times 64\times 20} tensor}
 \item{coeffiicients}{A \eqn{64\times 64 \times 1} tensor with the bat pattern}
 \item{Gamma}{Two envelope basis: \eqn{64 \times 14} matrices}
}}
\usage{
data("bat")
}
\description{
Simulated data used in tensor response regression (TRR). The pattern of coefficient is a bat.
}
\examples{
data("bat")
## Coefficients
coeff <- bat$coefficients
image(-coeff@data[, , 1], axes=TRUE, col = grey(seq(0, 1, length = 256)))
title('Coefficient matrix')
}
\keyword{datasets}
