% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcov.Tenv.R
\name{vcov.Tenv}
\alias{vcov.Tenv}
\title{Calculate standard error for Tenv object.}
\usage{
\method{vcov}{Tenv}(object, ...)
}
\arguments{
\item{object}{an object of class "Tenv", as from \code{\link{TPR}} or \code{\link{TRR}}.}

\item{...}{arguments to be passed to or from other methods.}
}
\value{
The standard error tensor is returned.
}
\description{
Returns the standard error for Tenv object returned from \code{\link{TRR}}. There is no \code{vcov} method for object returned
from \code{\link{TPR}}.
}
\examples{
data("bat")
Xn <- bat$Xn
Yn <- bat$Yn
fit <- TRR(Xn, Yn, method="standard")
vcov(fit)
}
\seealso{
\code{\link{summary.Tenv}}
}
